code fixes:  server handler for GetLatestVersions passes through "stages" argument
tests:
 - handler test for GetLatestVersions
 - integration test for GetLatestVersions
 - refactor integration tests
